backend/requirements.txt
asyncpg==0.29.0
fastapi==0.111.1
pytest==8.2.2
pytest-asyncio==0.23.8
ruff==0.5.3
mypy==1.10.1
python-jose==3.3.0
pre-commit==3.7.1
pydantic-settings==2.3.4
